requirements.txt
kokoro-onnx
onnxruntime-gpu
sounddevice
numpy
requests
tqdm
scipy
customtkinter
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-cuda-runtime-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nvjitlink-cu12
